Split Secrets & Webhooks into Vaults and Webhooks
Refactor UI to split the legacy "Secrets & Webhooks" page into two focused pages: Vaults (organization tab) and Webhooks (standalone page). Added new pages frontend/src/pages/Vaults.jsx and frontend/src/pages/Webhooks.jsx, removed frontend/src/pages/SecretsWebhooks.jsx, and updated routes in App.jsx to register /vaults and /webhooks while redirecting legacy /secrets routes to the new locations. Updated CommandPalette, organization tabs, and sidebar items to reflect Vaults/Webhooks and adjusted ADVANCED_ITEM_IDS. UI comments explain intent and migration. This keeps secret storage under Organization and makes webhooks a dedicated console.
