+++ frontend/src/components/CommandPalette.jsx
@@ -43,7 +43,8 @@ const STATIC_PAGES = [
     { label: 'Queue', path: '/queue', category: 'Pages', keywords: 'bus operations tasks' },
     { label: 'Jobs', path: '/jobs', category: 'Pages', keywords: 'scheduler background work' },
     { label: 'Telemetry', path: '/telemetry', category: 'Pages', keywords: 'metrics observability' },
-    { label: 'Secrets', path: '/secrets', category: 'Pages', keywords: 'webhooks tokens vault' },
+    { label: 'Vaults', path: '/vaults', category: 'Pages', keywords: 'secrets tokens credentials vault' },
+    { label: 'Webhooks', path: '/webhooks', category: 'Pages', keywords: 'webhook receive forward delivery' },
     { label: 'Settings', path: '/settings', category: 'Settings', keywords: 'profile preferences' },
     { label: 'Settings: Users', path: '/settings/users', category: 'Settings', keywords: 'accounts team' },
     { label: 'Settings: API Keys', path: '/settings/api', category: 'Settings', keywords: 'tokens access' },
